Alternate recipe that excludes configuration file
We are using this recipe where I work, but the config_plist variable in this is not needed and actually causing us the issue of having to delete it. Could you maybe offer this as a separate pkg recipe without that particular code? We like the inclusion of the version in the package name though. That part is very helpful.